feat : Add AbstractEntity.class
- 엔티티 클래스들의 공통 부분을 추출해 중복 코드를 제거
