+++ src/main/java/com/codessquad/qna/QnaApplication.java
@@ -2,8 +2,10 @@ package com.codessquad.qna;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class QnaApplication {
     public static void main(String[] args) {
         SpringApplication.run(QnaApplication.class, args);
