+++ src/main/java/com/codessquad/qna/web/domain/Answer.java
@@ -5,13 +5,7 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 @Entity
-public class Answer {
-    private static final DateTimeFormatter ANSWER_DATETIME_FORMAT = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
+public class Answer extends AbstractEntity {
     @ManyToOne
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_answer_writer"))
     private User writer;
@@ -22,7 +16,6 @@ public class Answer {
 
     @Column(length = 1000)
     private String contents;
-    private LocalDateTime createdDateTime;
 
     protected Answer() {
     }
@@ -31,11 +24,6 @@ public class Answer {
         this.writer = writer;
         this.question = question;
         this.contents = contents;
-        this.createdDateTime = LocalDateTime.now();
-    }
-
-    public long getId() {
-        return this.id;
     }
 
     public User getWriter() {
@@ -44,13 +32,6 @@ public class Answer {
 
     public String getContents() {
         return this.contents;
-    }
-
-    public String getFormattedCreatedDate() {
-        if (createdDateTime == null) {
-            return "";
-        }
-        return createdDateTime.format(ANSWER_DATETIME_FORMAT);
     }
 
     public boolean isSameWriter(User writer) {
