+++ src/main/java/com/codessquad/qna/web/domain/Question.java
@@ -3,19 +3,11 @@ package com.codessquad.qna.web.domain;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
-public class Question {
-    private static final DateTimeFormatter QUESTION_DATETIME_FORMAT = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
+public class Question extends AbstractEntity {
     @ManyToOne
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_questions_writer"))
     private User writer;
@@ -25,7 +17,6 @@ public class Question {
 
     @Column(length = 50000)
     private String contents;
-    private LocalDateTime createdDateTime = LocalDateTime.now();
 
     @OneToMany(mappedBy = "question")
     @JsonIgnore
@@ -46,14 +37,6 @@ public class Question {
         this.contents = contents;
     }
 
-    public void setCreatedDate() {
-        this.createdDateTime = LocalDateTime.now();
-    }
-
-    public long getId() {
-        return id;
-    }
-
     public User getWriter() {
         return writer;
     }
@@ -68,13 +51,6 @@ public class Question {
 
     public List<Answer> getAnswers() {
         return this.answers;
-    }
-
-    public String getFormattedCreatedDate() {
-        if (createdDateTime == null) {
-            return "";
-        }
-        return createdDateTime.format(QUESTION_DATETIME_FORMAT);
     }
 
     public int getCountOfAnswers() {
