+++ src/main/java/com/codessquad/qna/web/domain/User.java
@@ -2,15 +2,11 @@ package com.codessquad.qna.web.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import javax.persistence.*;
-import java.util.Objects;
+import javax.persistence.Column;
+import javax.persistence.Entity;
 
 @Entity
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
+public class User extends AbstractEntity {
     @Column(nullable = false, length = 20, unique = true)
     private String userId;
     @JsonIgnore
@@ -34,10 +30,6 @@ public class User {
         this.email = email;
     }
 
-    public Long getId() {
-        return id;
-    }
-
     public String getUserId() {
         return userId;
     }
@@ -58,26 +50,9 @@ public class User {
         return this.password.equals(testPassword);
     }
 
-    public boolean isSameId(long id) {
-        return this.id == id;
-    }
-
     public void update(User user) {
         this.password = user.password;
         this.name = user.name;
         this.email = user.email;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof User)) return false;
-        User user = (User) o;
-        return Objects.equals(getId(), user.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(getId());
     }
 }
