+++ src/main/resources/data.sql
@@ -1,6 +1,6 @@
-INSERT INTO USER (ID, USER_ID, PASSWORD, NAME, EMAIL) VALUES ('1', 'javajigi', 'test', '자바지기', 'javajigi@slipp.net')
-INSERT INTO USER (ID, USER_ID, PASSWORD, NAME, EMAIL) VALUES ('2', 'sanjigi', 'test', '산지기', 'sanjigi@slipp.net')
-INSERT INTO USER (ID, USER_ID, PASSWORD, NAME, EMAIL) VALUES ('3', 'hi', 'hi', 'hi', 'hi@slipp.net')
+INSERT INTO USER (ID, CREATED_DATE_TIME, USER_ID, PASSWORD, NAME, EMAIL) VALUES ('1', '2021-03-22 17:03:01.106461', 'javajigi', 'test', '자바지기', 'javajigi@slipp.net')
+INSERT INTO USER (ID, CREATED_DATE_TIME, USER_ID, PASSWORD, NAME, EMAIL) VALUES ('2', '2021-03-22 17:03:01.106461', 'sanjigi', 'test', '산지기', 'sanjigi@slipp.net')
+INSERT INTO USER (ID, CREATED_DATE_TIME, USER_ID, PASSWORD, NAME, EMAIL) VALUES ('3', '2021-03-22 17:03:01.106461', 'hi', 'hi', 'hi', 'hi@slipp.net')
 
 INSERT INTO QUESTION (ID, WRITER_ID, TITLE, CONTENTS, CREATED_DATE_TIME, COUNT_OF_ANSWERS) VALUES ('1', '3', 'hello', 'bye', '2021-03-22 17:03:01.106461', '0')
 INSERT INTO QUESTION (ID, WRITER_ID, TITLE, CONTENTS, CREATED_DATE_TIME, COUNT_OF_ANSWERS) VALUES ('2', '1', 'hoi', 'dui', '2021-03-20 16:03:03.106461', '1')
